Updated day end
The strange happenings in each scene will still happen even if the time
when it would happen has already passed. In other words if you were in
the classroom at 30 sec when the lockers are supposed to start shaking
then walk into the hallway the lockers will start shaking upon entering.
Before, those strange happenings would only happen if you were in the
room in question at the moment it's supposed to initially happen. I
don't know why I decided to write such a detailed comment for this
commit but I did so deal with it.